Comments: nits and small refactoring.
