+++ src/main/java/org/mozilla/gecko/sync/repositories/InactiveSessionException.java
@@ -43,9 +43,6 @@ public class InactiveSessionException extends SyncException {
 
   private static final long serialVersionUID = 537241160815940991L;
 
-  public InactiveSessionException() {
-    this(null);
-  }
   public InactiveSessionException(Exception ex) {
     super(ex);
   }
