+++ src/main/java/org/mozilla/gecko/sync/stage/PasswordsServerSyncStage.java
@@ -4,10 +4,12 @@
 
 package org.mozilla.gecko.sync.stage;
 
-import org.mozilla.gecko.sync.repositories.PasswordsRepository;
+import org.mozilla.gecko.sync.CryptoRecord;
 import org.mozilla.gecko.sync.repositories.RecordFactory;
 import org.mozilla.gecko.sync.repositories.Repository;
-import org.mozilla.gecko.sync.repositories.domain.PasswordRecordFactory;
+import org.mozilla.gecko.sync.repositories.android.PasswordsRepositorySession;
+import org.mozilla.gecko.sync.repositories.domain.PasswordRecord;
+import org.mozilla.gecko.sync.repositories.domain.Record;
 
 public class PasswordsServerSyncStage extends ServerSyncStage {
 
@@ -22,12 +24,22 @@ public class PasswordsServerSyncStage extends ServerSyncStage {
 
   @Override
   protected Repository getLocalRepository() {
-    return new PasswordsRepository();
+    return new PasswordsRepositorySession.PasswordsRepository();
   }
 
   @Override
   protected RecordFactory getRecordFactory() {
     return new PasswordRecordFactory();
+  }
+
+  public class PasswordRecordFactory extends RecordFactory {
+
+    @Override
+    public Record createRecord(Record record) {
+      PasswordRecord r = new PasswordRecord();
+      r.initFromEnvelope((CryptoRecord) record);
+      return r;
+    }
   }
 
 }
